double_codification/requirements.txt
Django==5.1.5
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
crispy-tailwind==1.0.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dj-rest-auth==7.0.1
django-cors-headers==4.6.0
psycopg==3.2.3
psycopg-binary==3.2.3
django-filter==24.3
Markdown==3.7
drf-yasg==1.21.8
pandas==2.2.3
django-tables2
uvicorn==0.34.0
gunicorn==23.0.0
django-tenants==3.7.0
python-decouple==3.8
Unidecode==1.3.8
django-auth-ldap==5.1.0
